// src/store/api/leadsApi.ts
export type LeadHistoryNoteType = {
  uid: string;
  leadUid: string;
  clientUid: string | null;
  note: string;
  historyNoteType: string;
  changeIa: boolean;
  ownerUid: string;
  createdAt: string;
  updatedAt: string;
  owner?: {
    uid: string;
    login: string;
    profile?: {
      uid: string;
      name: string;
      email: string;
      phone: string | null;
      avatar: string | null;
    };
  };
};